+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+### Fixed
+* [Demo]Fixed the remote video render delay
+
 ## [0.17.9] - 2022-12-02
 
 ## [0.17.8] - 2022-11-16